requirements.txt
# =============================================================================
# Oprina Requirements - Clean ADK Setup
# =============================================================================

# Core ADK (Agent Development Kit)
google-adk

# Google AI/Cloud (minimal set to avoid conflicts)
google-generativeai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# LLM
litellm

# Database and Storage
supabase
asyncpg
psycopg2-binary

# Web Framework and API
fastapi
uvicorn
pydantic
pydantic-settings

# Voice Processing
speechrecognition
pyaudio
pydub
websockets

# HTTP and Async
httpx
aiohttp
aiofiles
python-multipart

# Utilities and Data Processing
python-dotenv
python-jose
passlib
email-validator
python-dateutil
pytz

# Logging
structlog
colorama

# Development and Testing
pytest
pytest-asyncio
pytest-cov
black
isort
mypy

# Core dependencies
python-dotenv==1.0.0
pydantic==2.4.2
fastapi==0.103.1
uvicorn==0.23.2
websockets==11.0.3
asyncio==3.4.3
aiohttp==3.8.5
requests==2.31.0

# Google API dependencies
google-api-python-client==2.97.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

# ADK dependencies
# Note: These are placeholders. Replace with actual ADK dependencies when available.
# adk==0.1.0
# adk-tools==0.1.0

# Logging and monitoring
logging==0.5.1.2
structlog==23.1.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1